sql quiries separated from controllers to services in backend
