+++ backend/controllers/commonController.js
@@ -1,5 +1,6 @@
 const db = require("../config/db");
 const { getUserFromCookies } = require('../middlewares/authMiddleware');
+const commonService = require('../services/commonService');
 
 //get all counts of cases, complaints, investigations, and evidences
 exports.getAllStatsCount = async (req, res) => {
@@ -9,48 +10,23 @@ exports.getAllStatsCount = async (req, res) => {
             return res.status(401).json({ message: "No token provided" });
         }
 
-        const user  = await getUserFromCookies(token);
+        const user = await getUserFromCookies(token);
         if (!user) {
             return res.status(401).json({ message: "Unauthorized" });
         }
 
-        let query
-        if (user.role === "OIC" || user.role === "Crime OIC") {
-            query = `
-                SELECT
-                    (SELECT COUNT(*) FROM cases WHERE status = 'inprogress') AS total_active_cases,
-                    (SELECT COUNT(*) FROM complaints WHERE status = 'new') AS total_new_complaints,
-                    (SELECT COUNT(*) FROM investigation WHERE status = 'inprogress') AS total_active_investigations
-                `;
-        }
-        else if (user.role === "Sub Inspector") {
-            query = `
-                SELECT
-                    (SELECT COUNT(*) FROM cases WHERE status = 'inprogress' AND leader_id = ?) AS total_active_cases,
-                    (SELECT COUNT(*) FROM complaints WHERE status = 'new') AS total_new_complaints,
-                    (SELECT COUNT(*) FROM investigation WHERE status = 'inprogress' AND leader_id = ?) AS total_active_investigations
-                `;
-        }
-        else if (user.role === "Sergeant") {
-            query = `
-                SELECT
-                    (SELECT COUNT(*) FROM cases WHERE status = 'inprogress' AND leader_id = ?) AS total_active_cases,
-                    (SELECT COUNT(*) FROM complaints WHERE status = 'new') AS total_new_complaints,
-                    (SELECT COUNT(*) FROM investigation WHERE status = 'inprogress' AND leader_id = ?) AS total_active_investigations
-                `;
-        }
-        else {
+        if (!["OIC", "Crime OIC", "Sub Inspector", "Sergeant"].includes(user.role)) {
             return res.status(403).json({ message: "Forbidden" });
         }
 
-        const [rows] = await db.query(query);
+        const stats = await commonService.getAllStatsCount(user.role, user.id);
         
-        if (rows.length === 0) {
+        if (!stats) {
             console.log("No stats found");
             return res.status(404).json({ message: "No stats found" });
         }
         
-        res.status(200).json({ statsValues: rows[0] });
+        res.status(200).json({ statsValues: stats });
     } catch (error) {
         console.error("Error fetching stats:", error);
         res.status(500).json({ message: "Internal server error" });
